Directory fix again
